Combine all integration components into a single adapter
