+++ libraries/botframework-streaming-extensions/src/Integration/index.ts
@@ -1,3 +1,1 @@
-export * from './NamedPipeConnector';
-export * from './StreamingRequestHandler';
-export * from './WebSocketConnector';
+export * from './BotFrameworkStreamingAdapter';
